@@ -3,11 +3,11 @@
 
 IntPoint::IntPoint(int x, int y) : x(x), y(y) {}
 
-int IntPoint::getX() {
+int IntPoint::getX() const {
     return x;
 }
 
-int IntPoint::getY() {
+int IntPoint::getY() const {
     return y;
 }
 
@@ -15,7 +15,7 @@ IntPoint::operator std::string() const {
     return "(" + std::to_string(x) + "," + std::to_string(y) + ")";
 }
 
-double IntPoint::getAngleTo(IntPoint &point2) {
+double IntPoint::getAngleTo(const IntPoint &point2) const {
     int x2 = point2.x;
     int y2 = point2.y;
     return abs((y2 - y) / (double) (x2 - x));
